removing output files (but keep them in tarball)
The man files are generated, but doing so requires a lot
of dependencies. So we remove the files from git to keep
out of git what does not belong there, but make sure we still
distribute them. Thus, developers are required to set up the
necessary environment, while end-users just wanting to compile
from source do not need them.