removed mutex calls from dbgprint, restructured
These calls were a real problem in regard to valgrind's threading
error detector (as well as strace). Alos, the limited parllel processing
under debub mode, which made it harder to reproduce some problems. The new
code is, I think, a good compromise between correctness and not affecting
the "real" code.