bugfix: race condition during directory creation
If multiple files try to create a directory at (almost) the same time,
some of them may fail. This is a data race. It likely can trigger another
error condition (yet unknown) that then causes a segfault.
The current fix (and probably the final one) is to make this operation
per-process exclusive. That should reduce concurrency in an acceptable way.