Add inotify support to imfile
Add "on file opened" and "on file closed" callbacks to strm_t that we
can setup in imfile. We then use them to setup inotify watches on
files we monitor. That way, imfile doesn't have to poll.

The patch probably needs some further work but the basic use case
seems to work fine.

This patch is rebased on the v5-stable commit:
536c9370f0 - Fix for "/run/systemd/journal/syslog" detection